+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add documentation for n-grams and skip n-grams
 * Add codecov
 * Add tidiers for LDA objects from topicmodels
+* Fixed a bug when tidying by line/sentence/paragraph/regex and there are multiple non-text columns
 
 # tidytext 0.1.0